requirements.txt
google-cloud-discoveryengine>=0.11.0
python-dotenv>=1.0.0
